cloud-function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage>=1.42.3
requests>=2.26.0
